+++ lib/crossbeams/layout/renderer/input.rb
@@ -17,7 +17,7 @@ module Crossbeams
           date_related_value_getter
 
           <<-HTML
-          <div #{wrapper_id} class="#{div_class}">#{hint_text}#{copy_prefix}
+          <div #{wrapper_id} class="#{div_class}"#{wrapper_visibility}>#{hint_text}#{copy_prefix}
             <input type="#{input_type}" value="#{CGI.escapeHTML(value.to_s)}" #{name_attribute} #{field_id} #{attr_list(datalist).join(' ')}>#{copy_suffix}
             <label for="#{@page_config.name}_#{@field_name}">#{@caption}#{error_state}#{hint_trigger}</label>
             #{datalist}